Changed floating point support to use Float class
This class can represent floats as fixed point 32 or 64 bit integers or IEEE floats or doubles.
Float32 produces smaller code that FloatDouble but only by about 500 bytes. I will continue to
optimize to get that number smaller. I also want to make a version where floats are not supported
at all, making some sort of dummy Float class

This gets the bootloader working again. I think the problem was that the presence of double parameters
were crashing because I was never initializing the floating point unit.

+++ baremetal/Serial.cpp
@@ -93,7 +93,7 @@ Serial::Error Serial::puts(const char* s, uint32_t size)
 	return Error::OK;
 }
 
-Serial::Error Serial::puts(double v)
+Serial::Error Serial::puts(Float v)
 {
     char buf[Print::MaxToStringBufferSize];
     if(Print::toString(buf, v)) {
